requirements.txt
httpx==0.27.2
python-multipart==0.0.9
google-cloud-storage==2.18.2
starlette==0.38.6
